backend/requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
Pillow
opencv-python
numpy
